chore: movie route added

+++ src/controllers/MovieController.ts
@@ -1,4 +1,6 @@
 import { type Request, type Response } from 'express'
+import { validationResult } from 'express-validator'
+
 import { MovieService } from '../services/MovieService'
 import { MovieRepository } from '../repositories/MovieRepository'
 import { type MovieInterface } from '../interfaces/MovieInterface'
@@ -7,33 +9,64 @@ const movieRepository = new MovieRepository()
 const movieService = new MovieService(movieRepository)
 
 export class MovieController {
-  async index(req: Request, res: Response): Promise<Response> {
+  async index(req: Request, res: Response): Promise<void> {
     const movies = await movieService.findAll()
-    return res.json(movies)
+    res.json(movies)
   }
 
-  async show(req: Request, res: Response): Promise<Response> {
+  async show(req: Request, res: Response): Promise<void> {
+    const errors = validationResult(req)
+
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() })
+      return
+    }
+
     const { id } = req.params
     const movie = await movieService.findById(Number(id))
-    return res.json(movie)
+
+    res.json(movie)
   }
 
-  async store(req: Request, res: Response): Promise<Response> {
+  async store(req: Request, res: Response): Promise<void> {
+    const errors = validationResult(req)
+
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() })
+      return
+    }
+
     const movie = req.body as MovieInterface
 
     const newMovie = await movieService.create(movie)
-    return res.json(newMovie)
+    res.status(201).json(newMovie)
   }
 
-  async update(req: Request, res: Response): Promise<Response> {
-    const movie = req.body
+  async update(req: Request, res: Response): Promise<void> {
+    const errors = validationResult(req)
+
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() })
+      return
+    }
+
+    const movie = req.body as MovieInterface
     const updatedMovie = await movieService.update(movie)
-    return res.json(updatedMovie)
+
+    res.json(updatedMovie)
   }
 
-  async destroy(req: Request, res: Response): Promise<Response> {
+  async destroy(req: Request, res: Response): Promise<void> {
+    const errors = validationResult(req)
+
+    if (!errors.isEmpty()) {
+      res.status(400).json({ errors: errors.array() })
+      return
+    }
+
     const { id } = req.params
-    const deletedMovieId = await movieService.delete(Number(id))
-    return res.json(deletedMovieId)
+    await movieService.delete(Number(id))
+
+    res.status(204).end()
   }
 }
